Add rib error
