+++ golem-rib/src/expr.rs
@@ -16,7 +16,6 @@ use crate::call_type::{CallType, InstanceCreationType};
 use crate::generic_type_parameter::GenericTypeParameter;
 use crate::parser::block::block;
 use crate::parser::type_name::TypeName;
-use crate::rib_compilation_error::RibCompilationError;
 use crate::rib_source_span::SourceSpan;
 use crate::type_registry::FunctionTypeRegistry;
 use crate::{
